Refine effects based on optimizer-derived information
The optimizer may be able to derive information that is not available
to inference. For example, it may SROA a mutable value to derive
additional constant information. Additionally, some effects, like
:consistent are path-dependent and should ideally be scanned once
all optimizations are done. Now, there is a bit of a complication
that we have generally so far taken the position that the optimizer
may do non-IPO-safe optimizations, although in practice we never
actually implemented any. This was a sensible choice, because
we weren't really doing anything with the post-optimized IR other
than feeding it into codegen anyway. However, with irinterp and
this change, there's now two consumers of IPO-safely optimized
IR. I do still think we may at some point want to run passes
that allow IPO-unsafe optimizations, but we can always add them
at the end of the pipeline.

With these changes, the effect analysis is a lot more precise.
For example, we can now derive :consistent for these functions:
```
function f1(b)
    if Base.inferencebarrier(b)
        error()
    end
    return b
end

function f3(x)
    @fastmath sqrt(x)
    return x
end
```
and we can derive `:nothrow` for this function:
```
function f2()
    if Ref(false)[]
        error()
    end
    return true
end
```